requirements.txt
matplotlib
numpy
pandas
scikit
learn
TA-Lib
tensorflow
addons
tensorflow
yfinance
